Fuses ContainerDescriptor and VisibleContainer
