--- Assets/Engine/Inventory/VisibleContainer.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: e3ca34599d24491183415a2fdc4fc033
-timeCreated: 1605024496
